+++ lib/core.dart
@@ -4,4 +4,5 @@ export 'package:actions_toolkit_dart/src/core/exit_code.dart';
 export 'package:actions_toolkit_dart/src/core/input_options.dart';
 export 'package:actions_toolkit_dart/src/core/models/summary_image_options.dart';
 export 'package:actions_toolkit_dart/src/core/models/summary_table_cell.dart';
+export 'package:actions_toolkit_dart/src/core/path_utils.dart';
 export 'package:actions_toolkit_dart/src/core/summary.dart';